+++ app/src/main/java/com/termux/app/datastore/termux/TerminalIOPreferencesDataStore.java
@@ -1,0 +1,61 @@
+package com.termux.app.datastore.termux;
+
+import android.content.Context;
+
+import androidx.preference.PreferenceDataStore;
+
+import com.termux.shared.termux.settings.preferences.TermuxAppSharedPreferences;
+
+public class TerminalIOPreferencesDataStore extends PreferenceDataStore {
+
+    private final Context mContext;
+    private final TermuxAppSharedPreferences mPreferences;
+
+    private static TerminalIOPreferencesDataStore mInstance;
+
+    private TerminalIOPreferencesDataStore(Context context) {
+        mContext = context;
+        mPreferences = TermuxAppSharedPreferences.build(context, true);
+    }
+
+    public static synchronized TerminalIOPreferencesDataStore getInstance(Context context) {
+        if (mInstance == null) {
+            mInstance = new TerminalIOPreferencesDataStore(context);
+        }
+        return mInstance;
+    }
+
+
+
+    @Override
+    public void putBoolean(String key, boolean value) {
+        if (mPreferences == null) return;
+        if (key == null) return;
+
+        switch (key) {
+            case "soft_keyboard_enabled":
+                mPreferences.setSoftKeyboardEnabled(value);
+                break;
+            case "soft_keyboard_enabled_only_if_no_hardware":
+                mPreferences.setSoftKeyboardEnabledOnlyIfNoHardware(value);
+                break;
+            default:
+                break;
+        }
+    }
+
+    @Override
+    public boolean getBoolean(String key, boolean defValue) {
+        if (mPreferences == null) return false;
+
+        switch (key) {
+            case "soft_keyboard_enabled":
+                return mPreferences.isSoftKeyboardEnabled();
+            case "soft_keyboard_enabled_only_if_no_hardware":
+                return mPreferences.isSoftKeyboardEnabledOnlyIfNoHardware();
+            default:
+                return false;
+        }
+    }
+
+}
