+++ app/src/main/java/com/termux/app/fragments/settings/termux/TerminalIOPreferencesFragment.java
@@ -4,12 +4,11 @@ import android.content.Context;
 import android.os.Bundle;
 
 import androidx.annotation.Keep;
-import androidx.preference.PreferenceDataStore;
 import androidx.preference.PreferenceFragmentCompat;
 import androidx.preference.PreferenceManager;
 
 import com.termux.R;
-import com.termux.shared.termux.settings.preferences.TermuxAppSharedPreferences;
+import com.termux.app.datastore.termux.TerminalIOPreferencesDataStore;
 
 @Keep
 public class TerminalIOPreferencesFragment extends PreferenceFragmentCompat {
@@ -27,56 +26,4 @@ public class TerminalIOPreferencesFragment extends PreferenceFragmentCompat {
 
 }
 
-class TerminalIOPreferencesDataStore extends PreferenceDataStore {
 
-    private final Context mContext;
-    private final TermuxAppSharedPreferences mPreferences;
-
-    private static TerminalIOPreferencesDataStore mInstance;
-
-    private TerminalIOPreferencesDataStore(Context context) {
-        mContext = context;
-        mPreferences = TermuxAppSharedPreferences.build(context, true);
-    }
-
-    public static synchronized TerminalIOPreferencesDataStore getInstance(Context context) {
-        if (mInstance == null) {
-            mInstance = new TerminalIOPreferencesDataStore(context);
-        }
-        return mInstance;
-    }
-
-
-
-    @Override
-    public void putBoolean(String key, boolean value) {
-        if (mPreferences == null) return;
-        if (key == null) return;
-
-        switch (key) {
-            case "soft_keyboard_enabled":
-                    mPreferences.setSoftKeyboardEnabled(value);
-                break;
-            case "soft_keyboard_enabled_only_if_no_hardware":
-                mPreferences.setSoftKeyboardEnabledOnlyIfNoHardware(value);
-                break;
-            default:
-                break;
-        }
-    }
-
-    @Override
-    public boolean getBoolean(String key, boolean defValue) {
-        if (mPreferences == null) return false;
-
-        switch (key) {
-            case "soft_keyboard_enabled":
-                return mPreferences.isSoftKeyboardEnabled();
-            case "soft_keyboard_enabled_only_if_no_hardware":
-                return mPreferences.isSoftKeyboardEnabledOnlyIfNoHardware();
-            default:
-                return false;
-        }
-    }
-
-}
